euler 35

+++ src/test/kotlin/functions/IsPrimeTest.kt
@@ -1,16 +1,17 @@
 package functions
 
+import java.math.BigInteger
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class IsPrimeTest {
     @Test
     fun isPrimeTest() {
-        assertEquals(true, isPrime(7))
-        assertEquals(false, isPrime(25))
-        assertEquals(false, isPrime(14))
-        assertEquals(false, isPrime(4))
-        assertEquals(false, isPrime(1))
+        assertEquals(true, isPrime(BigInteger.valueOf(7)))
+        assertEquals(false, isPrime(BigInteger.valueOf(25)))
+        assertEquals(false, isPrime(BigInteger.valueOf(14)))
+        assertEquals(false, isPrime(BigInteger.valueOf(4)))
+        assertEquals(false, isPrime(BigInteger.valueOf(1)))
     }
 
 }
